separate payload types
